added fileUpload alternate submissions methods support
