+++ package.json
@@ -24,12 +24,12 @@
     "license": "MIT",
     "dependencies": {
         "axios": "0.21.4",
-        "camelcase": "6.2.0",
         "crypto-js": "4.0.0",
-        "file-type": "^16.5.4",
-        "form-data": "4.0.0",
         "joi": "17.4.0",
-        "query-string": "6.14.1"
+        "query-string": "6.14.1",
+        "camelcase": "6.2.0",
+        "form-data": "4.0.0",
+        "file-type": "^16.5.4"
     },
     "devDependencies": {
         "axios-mock-adapter": "1.19.0",
