+++ sdk/platform/PlatformAPIClient.js
@@ -25,7 +25,7 @@ class APIClient {
                 Authorization: "Bearer " + token,
                 ...headers,
             },
-            maxBodyLength: Infinity
+            maxBodyLength: Infinity,
         };
 
         return pdkAxios.request(rawRequest);
