+++ tests/platform/PixelbinClient.spec.js
@@ -142,7 +142,7 @@ describe("Pixelbin Client", () => {
             headers: {
                 Authorization: `Bearer ${Buffer.from("test-api-secret").toString("base64")}`,
             },
-            "maxBodyLength": Infinity,
+            maxBodyLength: Infinity,
         });
         requestMock.mockRestore();
     });
@@ -181,7 +181,7 @@ describe("Pixelbin Client", () => {
             headers: {
                 Authorization: `Bearer ${Buffer.from("test-api-secret").toString("base64")}`,
             },
-            "maxBodyLength": Infinity,
+            maxBodyLength: Infinity,
         });
         requestMock.mockRestore();
     });
@@ -213,7 +213,7 @@ describe("Pixelbin Client", () => {
                 Authorization: `Bearer ${Buffer.from("test-api-secret").toString("base64")}`,
                 "content-type": expect.stringMatching(/^multipart\/form-data; boundary=.*/),
             },
-            "maxBodyLength": Infinity,
+            maxBodyLength: Infinity,
         });
         expect(requestMock.mock.calls[0][0].data).toBeInstanceOf(FormData);
         requestMock.mockRestore();
@@ -246,7 +246,7 @@ describe("Pixelbin Client", () => {
             headers: {
                 Authorization: `Bearer ${Buffer.from("test-api-secret").toString("base64")}`,
             },
-            "maxBodyLength": Infinity,
+            maxBodyLength: Infinity,
         });
         expect(requestMock.mock.calls[0][0].data).toEqual({
             name: "New Name",
